refactor(parsers): move the diff section behind the parser seam (#148)
The comment formatter scanned raw CLI output for the '✓ Generated'
marker and sliced the text itself — parser work inside a formatter. The
same algorithm existed in the parser base class, where the diff parser
computed its result and threw it away.

The result now carries the section: the parser populates diffSection,
the formatter reads it. One implementation remains, owned by the
parsers, and the formatter no longer imports the pattern library at all.

The two implementations disagreed, so the choice is explicit. On no
match the parser's version returned the entire original output and the
formatter's returned an empty string. The formatter's behaviour is the
one that survives, because the other fallback would render a whole raw
CLI capture into a pull request comment where nothing renders today.
That disagreement is also why the discarded call was harmless.

With that fallback preserved the rendered text is unchanged, and the
committed snapshots are untouched, which is the behaviour-preservation
check.

The real-world diff fixture in the formatter tests is now built by
running the parser rather than by hand. A hand-written result would have
had to carry diffSection itself, which is the same drift #146 found in
the deploy capture.

Coverage moves within threshold, no re-baseline: 90.62 to 90.54
statements, 80.97 to 80.80 branches, 90.58 to 90.50 lines; functions
unmoved at 97.35.
