+++ __tests__/github/client.test.ts
@@ -182,6 +182,7 @@ describe("GitHub Client - API Integration", () => {
       changeSummary: "No infrastructure changes detected",
       changes: [],
       completionStatus: "complete",
+      diffSection: "",
       exitCode: 0,
       operation: "diff",
       plannedChanges: 0,
@@ -325,6 +326,7 @@ describe("GitHub Client - API Integration", () => {
           { action: "update", name: "Bucket1", type: "S3" },
         ],
         completionStatus: "complete",
+        diffSection: "",
         exitCode: 0,
         operation: "diff",
         plannedChanges: 5,
