+++ __tests__/operations/diff.test.ts
@@ -259,6 +259,7 @@ Changes detected in infrastructure.`,
       changeSummary: "Failed to execute SST diff command",
       changes: [],
       completionStatus: "failed",
+      diffSection: "",
       error: "Authentication failed: Invalid SST token",
       exitCode: -1,
       operation: "diff",
