+++ __tests__/outputs/formatter.test.ts
@@ -112,6 +112,7 @@ describe("Output Formatter - GitHub Actions Output Processing", () => {
             },
           ],
           completionStatus: "complete",
+          diffSection: "",
           exitCode: 0,
           operation: "diff",
           plannedChanges: 2,
@@ -153,6 +154,7 @@ describe("Output Formatter - GitHub Actions Output Processing", () => {
           changeSummary: "No changes detected",
           changes: [],
           completionStatus: "complete",
+          diffSection: "",
           exitCode: 0,
           operation: "diff",
           plannedChanges: 0,
@@ -401,6 +403,7 @@ describe("Output Formatter - GitHub Actions Output Processing", () => {
           changeSummary: "Found changes",
           changes: [],
           completionStatus: "complete",
+          diffSection: "",
           exitCode: 0,
           operation: "diff",
           plannedChanges: 5,
