+++ __tests__/outputs/integration.test.ts
@@ -65,6 +65,7 @@ describe("OutputFormatter Integration", () => {
           { action: "delete", details: "", name: "assets", type: "Bucket" },
         ],
         completionStatus: "complete",
+        diffSection: "",
         exitCode: 0,
         operation: "diff",
         plannedChanges: 3,
@@ -217,6 +218,7 @@ describe("OutputFormatter Integration", () => {
         changeSummary: "",
         changes: [],
         completionStatus: "complete",
+        diffSection: "",
         exitCode: 0,
         operation: "diff",
         plannedChanges: 1,
