+++ src/github/formatters.ts
@@ -3,7 +3,6 @@
  * Provides operation-specific formatting with rich markdown support
  */
 
-import { SSTPatterns } from "../parsers/patterns.js";
 import type {
   BaseOperationResult,
   DeployResult,
@@ -468,8 +467,8 @@ No infrastructure changes detected for this operation.`;
       return "No changes detected";
     }
 
-    // Extract the actual diff section from the raw output
-    const diffContent = this.extractDiffSection(result.rawOutput);
+    // The parser extracted this; the formatter no longer scans raw CLI output.
+    const diffContent = result.diffSection;
 
     if (!diffContent || diffContent.trim() === "") {
       // Fallback to simple summary if diff section extraction fails
@@ -489,37 +488,6 @@ No infrastructure changes detected for this operation.`;
     }
 
     return diffContent;
-  }
-
-  /**
-   * Extract the diff section from SST raw output
-   */
-  private extractDiffSection(rawOutput: string): string {
-    const lines = rawOutput.split("\n");
-    let diffStartIndex = -1;
-
-    // Find the "✓ Generated" marker
-    for (let i = 0; i < lines.length; i += 1) {
-      const line = lines[i];
-      if (line && SSTPatterns.sections.generated.test(line)) {
-        diffStartIndex = i + 1;
-        break;
-      }
-    }
-
-    if (diffStartIndex === -1 || diffStartIndex >= lines.length) {
-      return "";
-    }
-
-    // Get all lines after the "✓ Generated" marker
-    const diffLines = lines.slice(diffStartIndex);
-
-    // Remove any trailing empty lines
-    while (diffLines.length > 0 && diffLines.at(-1)?.trim() === "") {
-      diffLines.pop();
-    }
-
-    return diffLines.join("\n").trim();
   }
 
   /**
